HACK: to make sure amd started after xorg server
As said in previous commits, we have to move back ac.service from user session
to system. However since ac.service depends on xorg, and xorg is in user
session, we have to use *busy* checking /tmp/.X0-lock before we starts. This
is ugly, but at least a reasonable fix for now.

(If anyone has better solution, please go ahead to fix this.)

Also, both launchpad-preload@.service and ac.service are actually part of
graphical.target, not multi-user.target.

Change-Id: I37703578871c507d2d4c3c6d9e705e7784109396
Signed-off-by: Yin Kangkai <kangkai.yin@intel.com>