fix __app_dbus_signal_filter() not to fail on a valid unix user
When the AUL clients received a DBus app_launch or app_dead signal
the __app_dbus_signal_filter() function, that handles the message,
failed for every user ID that is not 0, ie. for all valid user ID's.

As a result the clients notification routines were not called and
the applications could not be tracked.

Fixes: TC-1514

Change-Id: Ib67ddfaebf62d26f01256c625d13d770e71f8d48