Run user application in PID namespace
If application run in namespace it would have different PID
in /proc and different returned by getpid(). So if application
is asking for it's own data from proc it is more securely to
read /proc/self/ then /proc/some_pid/.

This commit uses /proc/self/status file to read PID. This pid
is required to create socket that will be used in communication
with AUL (aul is working on host and does not know process pid).

Change-Id: Ie66d766943301f8df00ca4b83fba018bf28cb70c